Changes from #29 that were lost in merge...
All changes are from #29.  However, I didn't add @trishankatdatadog 's suggestion to add: "For an example of time estimates for implementation, please see [this thread](https://discuss.python.org/t/which-cryptographic-signing-approach/2241/5)."  I'd like to discuss this point.  I feel like we're not fully synced here and it feels like we're going off topic to point to this in our FAQ.